+++ src/scripts/index.js
@@ -2,31 +2,23 @@ import 'regenerator-runtime'; /* for async await transpile */
 import '../styles/main.css';
 import '../styles/universal.css';
 import '../styles/navigation.css';
+import '../styles/catalog.css';
+import '../styles/footer.css';
+import data from '../DATA.json';
+import navigation from './navigation';
+import restaurant from './restaurant';
 
 console.log('Hello Coders! :)');
 
-const navbar = document.querySelector("#nav");
 const navButtonOpen = document.querySelector("#navButtonOpen");
 const navButtonClose = document.querySelector("#navButtonClose");
 
-function openNav() {
-    // navbar.style.transform = "translateY(0%)";
-    navbar.classList.toggle('slide');
-    navButtonOpen.style.display = "none";
-    navButtonClose.style.display = "block";
-}
-
-function closeNav() {
-    navbar.classList.toggle('slide');
-    // navbar.style.transform = "translateY(-100%)";
-    navButtonOpen.style.display = "block";
-    navButtonClose.style.display = "none";
-}
-
 navButtonOpen.addEventListener('click', function () {
-    openNav();
+    navigation.openNav();
 });
 
 navButtonClose.addEventListener('click', function () {
-    closeNav();
+    navigation.closeNav();
 });
+
+restaurant.getRestaurant(data);
