+++ src/styles/catalog.css
@@ -1,0 +1,62 @@
+#catalog {
+    margin-top: 50px;
+    width: 100%;
+}
+
+#catalog .row {
+    display: grid;
+    grid-template-columns: auto auto auto;
+    margin: 0 auto;
+    width: 90%;
+    grid-row-gap: 1.5em;
+    height: auto;
+}
+
+#catalog .column {
+    flex-grow: 1;
+    margin: 15px;
+    cursor: pointer;
+    padding:15px;
+}
+
+#catalog .restaurant:hover {
+    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
+}
+
+#catalog .catalog-title {
+    font-size: 12pt;
+    font-weight: 800;
+    margin-top:10px;
+}
+
+#catalog .catalog-rating {
+    margin-top: 10px;
+    color: rgb(221, 221, 16);
+    font-weight: 800;
+    padding: 10px;
+}
+
+#catalog .header-text {
+    font-size: 22pt;
+    font-weight: 800;
+}
+
+#catalog .catalog-truncate {
+    overflow: hidden;
+    margin-top:10px;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+}
+
+@media (max-width: 768px) {
+    #catalog {
+        margin-top: 30px;
+    }
+
+    #catalog .row {
+        display: flex;
+        flex-direction: column;
+    }
+}
