+++ src/styles/universal.css
@@ -37,6 +37,26 @@ button:hover {
     transition: all 0.5s;
 }
 
+article {
+    max-width: 1296px;
+    margin: auto;
+}
+
+/* Skip link */
+.skip-link {
+    position: absolute;
+    top: -80px;
+    left: 0;
+    background-color: var(--btn);
+    color: white;
+    padding: 8px;
+    z-index: 10;
+}
+
+.skip-link:focus {
+    top: 0;
+}
+
 @media (max-width: 768px) {
     button {
         width: 100%;
