+++ src/templates/index.html
@@ -7,6 +7,7 @@
     <title>Restaurant Apps Starter Project</title>
 </head>
 <body>
+    <a href="#main-content" class="skip-link">Menuju ke konten</a>
     <header class="jumbotron">
         <nav>
             <div class="logoTextMobile">
@@ -21,15 +22,15 @@
                 </div>
                 <ul>
                     <li>
-                        <a class="active" href="#hero">HOME</a>
+                        <a class="active" href="/">HOME</a>
                     </li>
                     <li>
-                        <a href="#arrival">FAVORITE</a>
+                        <a href="#">FAVORITE</a>
                     </li>
                 </ul>
                 <ul>
                     <li>
-                        <a href="#about">ABOUT US</a>
+                        <a href="https://www.linkedin.com/in/gilangcsy/" target="_blank">ABOUT</a>
                     </li>
                 </ul>
             </div>
@@ -55,7 +56,23 @@
                 </div>
             </div>
         </section>
+        <article id="catalog">
+            <div class="row" id="main-content">
+                <div class="column">
+                    <div class="header-text">
+                        EXPLORE RESTAURANT
+                    </div>
+                </div>
+            </div>
+            <div class="row" id="restaurants">
+            </div>
+        </article>
     </main>
+    <footer id="footer">
+        <div class="row">
+            © 2021 KamengsFood. All Rights Reserved.
+        </div>
+    </footer>
 
 </body>
 </html>
